fix: Resolve major compilation errors during design system migration
- Add missing data classes (NFT, Collection, SendStatus) to NFTsViewModel
- Fix Transaction model conflicts by creating display-specific model
- Update imports to use correct BottomNavItem from components package
- Implement missing connect() method in WalletViewModel
- Fix getDynamicColors() calls to use LocalDynamicColors.current
- Add missing lazy grid items import for NFTsScreen
- Rename duplicate/backup files to .bak to exclude from compilation
- Fix enum value names (ERROR → FAILURE) in NFTsViewModel

+++ app/src/main/java/com/truetap/solana/seeker/ui/components/StandardComponents.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.truetap.solana.seeker.ui.components.layouts.TrueTapSpacing
 import com.truetap.solana.seeker.ui.components.layouts.TrueTapTypography
 import com.truetap.solana.seeker.ui.theme.*
@@ -385,7 +386,7 @@ fun TrueTapAvatar(
     ) {
         Text(
             text = text.take(2).uppercase(),
-            fontSize = (size * 0.4).sp,
+            fontSize = (size * 0.4).toInt().sp,
             fontWeight = FontWeight.Bold,
             color = dynamicColors.onPrimary,
             textAlign = TextAlign.Center
